+++ simple-stack/src/main/java/com/zhuinden/simplestack/ScopeLookupMode.java
@@ -1,0 +1,11 @@
+package com.zhuinden.simplestack;
+
+/**
+ * Specifies the mode used for looking up services within the scopes.
+ *
+ * Allows customizing whether both implicit and explicit parents are used, or only explicit parents.
+ */
+public enum ScopeLookupMode {
+    ALL,
+    EXPLICIT
+}