+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+-Simple Stack ?.??.? (????-??-??)
+--------------------------------
+
+- FIX: NPE when `canFindFromScope()` was used on an uninitialized stack, instead of returning `false`.
+
+- ADD: `ScopeLookupMode` for `canFindFromScope()` and `lookupFromScope()` methods, which allows restricting the lookup only to the explicit parent chain).
+
 -Simple Stack 1.13.2 (2019-02-05)
 --------------------------------